Analyze directory tree and archives
Add APIs to analyze a directory tree or an archive (zip, tar). By
default, all regular files with executable bit are analyzed except for
files that have a `.py`, `.sh`, or docs extension (md, rst, txt). In
addition, files with a pattern like `lib*.so`, `lib*.so.1.2.3`, and
`ld-*.so` are analyzed, too.

Signed-off-by: Christian Heimes <cheimes@redhat.com>